Reorganize files, add v3
